More sophisticated handling of --guess-* -- don't do it when running with -r
The eventual goal will be to have guessing happen by default on git,
and we never change the comment and description after the first
publish (the only one that doesn't have -r); after that we
auto-populate the per-publish comment instead.  But for now we only do
the first half of that.

In testing, noticed that --publish=false doesn't work, so added
--no-publish.