Refactor the guess-to-set-fields logic into its own routine.
When adding the code to add to change_description, I had only added it
in one place but it turns out I needed to add it to three.  "That's
ridiculous!" I said.  Hence the refactoring.

I have an error-exit in here so I can continue to test.  Will take
that out in a future commit.