Fix the logic that figures out the 'first' commit when guessing a summary.
When the user explicitly specifies a rev-range, it's easy: take the
first rev in the range.  Otherwise, we have to guess.  The right
guessing-point is the first commit *after* merge-base.  git isn't set
up to report on children like that, but I found a recipe here:
   http://stackoverflow.com/questions/1761825/referencing-the-child-of-a-commit-in-git/5353204#5353204

I also added a --no-merge flag to avoid the possibilities of having a
boring merge message be the summary.