Fix the code that guesses the description.
For some reason, I thought that every time you used post-review -r,
the description would only be the new commits since the last
post-review.  But that's not true: it has everything since you opened
the review.  So updating the description every time is the right thing
to do, and we don't need to mess with change_description at all.

Also, now that the summary is based on the *first* commit in the
review, and not the last, it will never change on subsequent
post-review calls, so it's safe to do --guess-summary even with
post-review -r (it may "update" the summary to be the exact same text
it was before).  So this simplifies the logic in the post-review
flag-parsing.