Add our own version of post-review, that runs our local version of postreview.py
The default post-review that comes with RBTools, and gets installed
into /usr/local/bin, runs the version of postreview.py that lives in
/usr/lib/python2.7 or the like.  We want to run the version from our
own repository.  This is a version of post-review that does that.

To use it, just add this bin directory to your path (before
/usr/local/bin and /usr/bin).

I take advantage of the fact we control this script, to also update
the code from git before running it -- that way everyone automagically
gets changes that we might make to reviewtools.  Since this is a bit
slow, I have code to only do the check once a day.