Add the ability to modify the last commit with reviewer info.
When this review request includes HEAD, and we're under git, we use
'git commit --amend' to modify the last commit and add a Reviewed-By
tag.  This tag includes the reviewers and the url of the review
request.

It is a TODO to update all the commits in a review request, not just
the last, and to support review requests that don't include HEAD.

Reviewed-By: tom,benkomalo <http://reviewboard.khanacademy.org/r/50/>