Add 'Reviewed-by' even when reviewers aren't specified: just say the review-url.
When people use 'post-review -r', to update an existing review,
they're not going to put a reviewer, so the old code wouldn't add a
Reviewed-by tag to the commit.  With this change, it does, just
specifying the review url (since the reviewers aren't known).  This is
still useful information to have.

Reviewed-By: benkomalo,tom <http://reviewboard.khanacademy.org/r/61/>