Use 'git notes' rather than 'git commit --amend' to update reviewed-by info.
This lets me resolve several TODOs, including modifying every commit
in a range, and modifying commits even when the revision-range is
specified explicitly in post-review.

The downside is that notes are nuked whenever someone runs 'git commit
--amend' (and maybe in other situations too, such as cherry-picking?)
but that's ok with me; I'm much more comfortable editing metadata like
this than actually amending a commit after it's been made.