Copy the reviewed-by message from previous commits when appropriate.
When post-review is run with -r, the first commits will have already
been updated with reviewer info.  In fact, they may even have better
information than we do, since they are more likely to have had the
reviewers specified.  So if we see an existing commit with reviewer
info, prefer that to our own reviewed-by test.