Revert "Don't crash when .hgrc has lines like 'key ='."
I can no longer reproduce this.  I've adeed some logging statements and the empty-string shouldn't be a problematic value:
['extensions.kilnauth', '/home/csilvers/khan/devtools/kilnauth/kilnauth.py\n']
['extensions.hgext.bookmarks', '\n']
['extensions.hggit', '\n']

Because newlines are included, the value should always be at least a
newline.  And indeed, when I revert to the old code, I don't get any
errors parsing the .hgrc.  I'm not sure how to reproduce the problem I
saw.  I'm reverting to be closer to upstream code.

This reverts commit c5cacb514daabe4c2c7f4409e7327e811d14fce3.